+++ packages/components-react/projects/nextjs/components/index.ts
@@ -1,2 +1,4 @@
+export * from './FooterPartials';
+export * from './HeaderPartials';
 export * from './TableWithCaption';
 export * from './Toast';