+++ packages/components/stencil.config.ts
@@ -68,6 +68,9 @@ export const config: Config = {
         preventAssignment: true,
         ROLLUP_REPLACE_IS_STAGING: isDevBuild ? '"staging"' : '"production"',
         ROLLUP_REPLACE_VERSION: `"${version}"`,
+        ROLLUP_REPLACE_CDN_BASE_URL: isDevBuild
+          ? '"http://localhost:3001"'
+          : 'document.porscheDesignSystem.cdn + "/porsche-design-system"', // document variable is set via components-js load() call
       }),
     ],
   },